Move Tab to its own file to remove Radium confusion
